% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilscidados-package.R
\docType{package}
\name{utilscidados-package}
\alias{utilscidados}
\alias{utilscidados-package}
\title{utilscidados: Internal Helpers for Insper Cidades Data Work}
\description{
The package supports two recurring workflows at Insper Cidades:
}
\details{
\enumerate{
\item Standardised data exports compliant with Insper's Dataverse
protocols, including file-format coverage, sanitised file names,
and column-level metadata (data dictionaries).
\item A simplified shapefile -> Mapbox Studio (GeoPortal) pipeline,
with automatic reprojection to WGS84 and consistent naming.
}
}
\section{Main functions}{

\itemize{
\item \code{\link[=export_table]{export_table()}} - write a data frame to csv / xlsx / parquet / rds / feather
\item \code{\link[=export_shapefile]{export_shapefile()}} - write an \code{sf} object to geojson / gpkg / shp / geoparquet
\item \code{\link[=export_tables_to_excel]{export_tables_to_excel()}} - write a list of data frames to a multi-sheet xlsx
\item \code{\link[=mapbox_upload]{mapbox_upload()}} - upload an \code{sf} object to Mapbox Studio as a tileset
\item \code{\link[=build_documentation]{build_documentation()}} - build a column-level data dictionary
\item \code{\link[=create_documentation_index]{create_documentation_index()}} - build an index of file/table names
}

Pass \code{extension = "dataverse"} to the exporters to select the
Dataverse-compliant subset of formats.
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/viniciusoike/utilscidados}
  \item Report bugs at \url{https://github.com/viniciusoike/utilscidados/issues}
}

}
\author{
\strong{Maintainer}: Vinicius Oike \email{viniciusor@insper.edu.br} [copyright holder]

}
\keyword{internal}
