% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilscidados-package.R
\docType{package}
\name{utilscidados-package}
\alias{utilscidados}
\alias{utilscidados-package}
\title{utilscidados: Internal Helpers for Insper Cidades Data Work}
\description{
The package supports two recurring workflows at Insper Cidades:
}
\details{
\enumerate{
\item Standardised data exports compliant with Insper's Dataverse
protocols, including file-format coverage, sanitised file names,
column-level metadata (data dictionaries), README skeletons, and
metadata workbook management.
\item A simplified shapefile -> Mapbox Studio (GeoPortal) pipeline,
with automatic reprojection to WGS84 and consistent naming.
}
}
\section{Main functions}{

\itemize{
\item \code{\link[=export_table]{export_table()}} - write a data frame to csv / xlsx / parquet / rds / feather
\item \code{\link[=export_shapefile]{export_shapefile()}} - write an \code{sf} object to geojson / gpkg / shp / geoparquet
\item \code{\link[=export_tables_to_excel]{export_tables_to_excel()}} - write a list of data frames to a multi-sheet xlsx
\item \code{\link[=mapbox_upload]{mapbox_upload()}} - upload an \code{sf} object to Mapbox Studio as a tileset
\item \code{\link[=build_documentation]{build_documentation()}} - build a column-level data dictionary
\item \code{\link[=create_documentation_index]{create_documentation_index()}} - build an index of file/table names
\item \code{\link[=build_readme]{build_readme()}} - generate a bilingual README.txt skeleton
\item \code{\link[=read_metadata]{read_metadata()}} / \code{\link[=update_metadata]{update_metadata()}} - read or write values in a
Dataverse metadata workbook
\item \code{\link[=create_metadata_sheet]{create_metadata_sheet()}} - clone a template sheet for a new dataset
\item \code{\link[=validate_metadata]{validate_metadata()}} - check a metadata sheet for missing fields or
placeholder markers
}

Pass \code{extension = "dataverse"} to the exporters to select the
Dataverse-compliant subset of formats.
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/portalcidados/utilscidados}
  \item Report bugs at \url{https://github.com/portalcidados/utilscidados/issues}
}

}
\author{
\strong{Maintainer}: Vinicius Oike \email{viniciusor@insper.edu.br} (\href{https://orcid.org/0009-0005-8015-9189}{ORCID}) [copyright holder]

}
\keyword{internal}
